fix: add hyperlinks to FAQ descriptions
