+++ frontend/src/pages/Home/Faqs.css
@@ -29,12 +29,9 @@
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
   padding: 15px 20px;
   cursor: pointer;
-  transition: background-color 0.3s ease;
+  transition: background-color 0.3s ease 0.5s, transform 0.5s ease;
 }
 
-.faq-item:hover {
-  background-color: #e6f7ff;
-}
 
 .faq-question {
   display: flex;
@@ -103,4 +100,16 @@
 
 .dark-mode .faq-question h3{
   color: #fff;
+}
+.faq-item:hover {
+  background-color: #e6f7ff;
+  transform: scale(1.02);
+}
+.faq-link {
+  color: #007bff;
+  text-decoration: underline;
+}
+
+.faq-link:hover {
+  text-decoration: none;
 }
